first level working, recusrion not working yet

+++ src/static_to_public.py
@@ -3,15 +3,15 @@
 # copy all files, subdirectories, nested files etc
 # log each copy
 from os import listdir, mkdir
-from os.path import exists, join, isfile
+from os.path import exists, join, isfile, isdir
 from shutil import copy, rmtree
 
 def copy_static_to_public(static, public):
     static_contents = find_dir_cont(static)
     public_contents = find_dir_cont(public)
     print(f"copy to dest public {public_contents[0]}, {public_contents[1]}, to deleted")
-    copy_to_dest(public_contents[0], public_contents[1], "deleted")
-    copy_to_dest(static_contents[0], static_contents[1], public)
+    copy_to_dest(public_contents[0], public_contents[1], public, "deleted")
+    copy_to_dest(static_contents[0], static_contents[1], static, public)
 
 
 def find_dir_cont(fromdir):
@@ -24,34 +24,36 @@ def find_dir_cont(fromdir):
         print(f"contents are : {items}")
         for item in items:
             diritem =  "/" + item
-            # print(f"[item] {item} isfile(diritem): {isfile(diritem)}")
-            if isfile(diritem):
+            fromdiritem = fromdir + "/" + item
+            print(f"[item] {item} isfile(fromdiritem): {isfile(fromdiritem)}, isdir(fromdiritem): {isdir(fromdiritem)}")
+            if not isdir(fromdiritem):
                 # pathdiritem = join(fromdir, diritem)
+                print(f"adding {diritem} to files")
                 files.append(diritem)
             else:
                 # pathdiritem = join(fromdir, diritem)
                 dirs.append(diritem)
+                print(f"[DIRECTORY FOUND] adding {diritem} to dirs, and searching further")
                 find_dir_cont(diritem)
     print(f"returing files {files}, and dirs {dirs}\n")
     return files, dirs
 
-def copy_to_dest(files, dirs,  todir):
+def copy_to_dest(files, dirs, fromdir, todir):
     print(f"[todir]: {todir}")
     print(f"[{todir} exists] {exists(todir)}")
+    if exists(todir) and todir != "deleted":
+        rmtree(todir)
     if not exists(todir):
         mkdir(todir)
-        print(f"[{todir} ] created {exists(todir)}")
-    if todir != "deleted":
-        rmtree(todir)
+        print(f"[{todir} created] {exists(todir)}")
 
     for d in dirs:
-        print(f"[copying d todir] d:{d} todir:{todir}")
+        print(f"[copying d todir] d:{d} todir: {todir}")
         fulld = todir + d
         if not exists(fulld):
             mkdir(fulld)
     for f in files:
-        print(f"[copting f todir] f:{f} todir:{todir}")
-        fullf = todir + f
-        if not exists(fullf):
-            copy(fullf, todir)
+        print(f"[copying f todir] f:{f} todir: {todir}")
+        fullf = fromdir + f
+        copy(fullf, todir)
     print(f"\nfiles and dirs copied to {todir}|\n**********\n")
